#170 Move the ArticlePreview component
